Assignment 40 partial

+++ app/controllers/posts_controller.rb
@@ -44,14 +44,15 @@ class PostsController < ApplicationController
   end
 
   def destroy
-    @post = Post.find(params[:id])
-
-    if @post.destroy
-      flash[:notice] = "\"#{@post.title}\" was deleted successfully."
-      redirect_to @post.topic
-    else
-      flash[:error] = "There was an error deleting the post."
-      render :show
+    if current_user.moderator? == false
+      @post = Post.find(params[:id])
+      if @post.destroy
+        flash[:notice] = "\"#{@post.title}\" was deleted successfully."
+        redirect_to @post.topic
+      else
+        flash[:error] = "There was an error deleting the post."
+        render :show
+      end
     end
   end
 
@@ -63,7 +64,7 @@ class PostsController < ApplicationController
 
   def authorize_user
     post = Post.find(params[:id])
-    unless current_user == post.user || current_user.admin?
+    unless current_user == post.user || current_user.admin? || current_user.moderator?
       flash[:error] = "You must be an admin to do that."
       redirect_to [post.topic, post]
     end
