+++ app/controllers/topics_controller.rb
@@ -15,13 +15,14 @@ class TopicsController < ApplicationController
   end
 
   def create
+    if current_user.moderator? == false
     @topic = Topic.new(topic_params)
-
-    if @topic.save
-      redirect_to @topic, notice: "Topic was saved successfully."
-    else
-      flash[:error] = "Error creating topic. Please try again."
-      render :new
+      if @topic.save
+        redirect_to @topic, notice: "Topic was saved successfully."
+      else
+        flash[:error] = "Error creating topic. Please try again."
+        render :new
+      end
     end
   end
 
@@ -43,14 +44,16 @@ class TopicsController < ApplicationController
   end
 
   def destroy
-    @topic = Topic.find(params[:id])
+    if current_user.moderator? == false
+      @topic = Topic.find(params[:id])
 
-    if @topic.destroy
-      flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
-      redirect_to action: :index
-    else
-      flash[:error] = "There was an error deleting the topic."
-      render :show
+      if @topic.destroy
+        flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
+        redirect_to action: :index
+      else
+        flash[:error] = "There was an error deleting the topic."
+        render :show
+      end
     end
   end
 
@@ -61,7 +64,7 @@ class TopicsController < ApplicationController
   end
 
   def authorize_user
-    unless current_user.admin?
+    unless current_user.admin? || current_user.moderator?
       flash[:error] = "You must be an admin to do that."
       redirect_to topics_path
     end
